# [REFACTOR] PromQL: DRY `stddev` and `stdvar` functions