This PR applies DRY principle to `histogram_stddev`, `histogram_stdvar`, `stddev_over_time` and to `stdvar_over_time`.

CC: @beorn7