@NeerajGartia21 2025-04-18
Thanks, @aknuds1, for the review. I've updated the code with your suggestions. Please take another look. The benchmark for the changes is below.

```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/promql
cpu: Apple M1
                                                                   │ /Users/neerajgartia/Desktop/dev/cp/before_1000.txt │ /Users/neerajgartia/Desktop/dev/cp/after_1000.txt │
                                                                   │                       sec/op                       │           sec/op            vs base               │
RangeQuery/expr=stddev_over_time(a_one[1d]),steps=1-8                                                      498.8µ ±  2%                 484.2µ ±  9%        ~ (p=0.132 n=6)
RangeQuery/expr=stddev_over_time(a_one[1d]),steps=100-8                                                    8.625m ±  1%                 9.055m ± 28%   +4.98% (p=0.002 n=6)
RangeQuery/expr=stddev_over_time(a_one[1d]),steps=1000-8                                                   82.46m ±  0%                 86.21m ±  9%   +4.54% (p=0.002 n=6)
RangeQuery/expr=stddev_over_time(a_ten[1d]),steps=1-8                                                      4.405m ±  5%                 4.323m ±  2%        ~ (p=0.394 n=6)
RangeQuery/expr=stddev_over_time(a_ten[1d]),steps=100-8                                                    84.48m ±  1%                 85.65m ±  1%   +1.39% (p=0.002 n=6)
RangeQuery/expr=stddev_over_time(a_ten[1d]),steps=1000-8                                                   818.6m ±  2%                 852.7m ± 46%   +4.17% (p=0.009 n=6)
RangeQuery/expr=stddev_over_time(a_hundred[1d]),steps=1-8                                                  42.64m ±  2%                 55.12m ± 39%  +29.25% (p=0.002 n=6)
RangeQuery/expr=stddev_over_time(a_hundred[1d]),steps=100-8                                                846.2m ±  0%                 887.8m ± 36%   +4.91% (p=0.002 n=6)
RangeQuery/expr=stddev_over_time(a_hundred[1d]),steps=1000-8                                                8.206 ±  1%                  8.584 ±  5%   +4.61% (p=0.002 n=6)
RangeQuery/expr=stdvar_over_time(a_one[1d]),steps=1-8                                                      489.9µ ±  1%                 488.5µ ± 10%        ~ (p=0.818 n=6)
RangeQuery/expr=stdvar_over_time(a_one[1d]),steps=100-8                                                    8.589m ±  1%                 8.768m ±  5%   +2.07% (p=0.002 n=6)
RangeQuery/expr=stdvar_over_time(a_one[1d]),steps=1000-8                                                   81.51m ±  1%                 84.47m ±  8%   +3.62% (p=0.002 n=6)
RangeQuery/expr=stdvar_over_time(a_ten[1d]),steps=1-8                                                      4.385m ±  7%                 4.381m ±  8%        ~ (p=0.937 n=6)
RangeQuery/expr=stdvar_over_time(a_ten[1d]),steps=100-8                                                    85.19m ±  1%                 86.35m ±  1%   +1.35% (p=0.002 n=6)
RangeQuery/expr=stdvar_over_time(a_ten[1d]),steps=1000-8                                                   828.5m ±  1%                 835.9m ±  8%        ~ (p=0.065 n=6)
RangeQuery/expr=stdvar_over_time(a_hundred[1d]),steps=1-8                                                  42.60m ±  1%                 43.54m ±  7%   +2.22% (p=0.002 n=6)
RangeQuery/expr=stdvar_over_time(a_hundred[1d]),steps=100-8                                                854.0m ±  8%                 901.8m ±  3%        ~ (p=0.065 n=6)
RangeQuery/expr=stdvar_over_time(a_hundred[1d]),steps=1000-8                                                8.435 ±  5%                  8.413 ±  1%        ~ (p=1.000 n=6)
RangeQuery/expr=histogram_stddev(rate(h_one[5m])),steps=1-8                                                49.19µ ± 25%                 39.69µ ±  4%  -19.33% (p=0.009 n=6)
RangeQuery/expr=histogram_stddev(rate(h_one[5m])),steps=100-8                                              174.4µ ± 19%                 159.7µ ±  2%   -8.41% (p=0.009 n=6)
RangeQuery/expr=histogram_stddev(rate(h_one[5m])),steps=1000-8                                             1.429m ± 21%                 1.160m ±  7%  -18.79% (p=0.026 n=6)
RangeQuery/expr=histogram_stddev(rate(h_ten[5m])),steps=1-8                                                406.8µ ± 51%                 299.0µ ±  3%  -26.50% (p=0.002 n=6)
RangeQuery/expr=histogram_stddev(rate(h_ten[5m])),steps=100-8                                              1.768m ± 36%                 1.451m ±  3%  -17.98% (p=0.002 n=6)
RangeQuery/expr=histogram_stddev(rate(h_ten[5m])),steps=1000-8                                             11.75m ± 30%                 11.95m ± 35%        ~ (p=0.589 n=6)
RangeQuery/expr=histogram_stddev(rate(h_hundred[5m])),steps=1-8                                            3.592m ± 55%                 3.185m ± 10%        ~ (p=0.394 n=6)
RangeQuery/expr=histogram_stddev(rate(h_hundred[5m])),steps=100-8                                          16.68m ± 45%                 15.91m ±  8%        ~ (p=0.180 n=6)
RangeQuery/expr=histogram_stddev(rate(h_hundred[5m])),steps=1000-8                                         116.2m ± 24%                 114.7m ±  5%        ~ (p=0.132 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_one[5m])),steps=1-8                                                43.79µ ± 17%                 39.36µ ±  0%  -10.11% (p=0.002 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_one[5m])),steps=100-8                                              215.6µ ± 26%                 159.3µ ±  1%  -26.14% (p=0.002 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_one[5m])),steps=1000-8                                             1.299m ± 39%                 1.145m ±  5%  -11.83% (p=0.004 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_ten[5m])),steps=1-8                                                376.0µ ± 29%                 294.7µ ±  1%  -21.62% (p=0.002 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_ten[5m])),steps=100-8                                              1.753m ± 23%                 1.439m ±  0%  -17.94% (p=0.002 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_ten[5m])),steps=1000-8                                             13.71m ± 34%                 10.81m ±  1%  -21.17% (p=0.002 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_hundred[5m])),steps=1-8                                            3.307m ± 31%                 2.902m ±  2%  -12.26% (p=0.002 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_hundred[5m])),steps=100-8                                          15.38m ±  2%                 14.65m ±  1%   -4.76% (p=0.002 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_hundred[5m])),steps=1000-8                                         117.3m ±  4%                 113.2m ±  2%   -3.49% (p=0.009 n=6)
geomean                                                                                                    11.88m                       11.24m         -5.44%

                                                                   │ /Users/neerajgartia/Desktop/dev/cp/before_1000.txt │ /Users/neerajgartia/Desktop/dev/cp/after_1000.txt │
                                                                   │                        B/op                        │            B/op              vs base              │
RangeQuery/expr=stddev_over_time(a_one[1d]),steps=1-8                                                    563.1Ki ±   1%                562.7Ki ±   1%       ~ (p=0.818 n=6)
RangeQuery/expr=stddev_over_time(a_one[1d]),steps=100-8                                                  554.1Ki ±   1%                555.1Ki ±   1%       ~ (p=0.310 n=6)
RangeQuery/expr=stddev_over_time(a_one[1d]),steps=1000-8                                                 588.4Ki ±   6%                589.8Ki ±   6%       ~ (p=0.584 n=6)
RangeQuery/expr=stddev_over_time(a_ten[1d]),steps=1-8                                                    754.0Ki ±   0%                755.0Ki ±   1%       ~ (p=0.699 n=6)
RangeQuery/expr=stddev_over_time(a_ten[1d]),steps=100-8                                                  789.6Ki ±   5%                789.6Ki ±   5%       ~ (p=0.712 n=6)
RangeQuery/expr=stddev_over_time(a_ten[1d]),steps=1000-8                                                1013.8Ki ±   0%                904.9Ki ±  14%       ~ (p=0.420 n=6)
RangeQuery/expr=stddev_over_time(a_hundred[1d]),steps=1-8                                                2.660Mi ±   1%                2.663Mi ±   2%       ~ (p=0.394 n=6)
RangeQuery/expr=stddev_over_time(a_hundred[1d]),steps=100-8                                              2.909Mi ±   8%                2.909Mi ±   8%       ~ (p=0.900 n=6)
RangeQuery/expr=stddev_over_time(a_hundred[1d]),steps=1000-8                                             3.337Mi ± 132%                3.335Mi ± 119%       ~ (p=0.394 n=6)
RangeQuery/expr=stdvar_over_time(a_one[1d]),steps=1-8                                                    560.1Ki ±   0%                560.7Ki ±   1%       ~ (p=0.485 n=6)
RangeQuery/expr=stdvar_over_time(a_one[1d]),steps=100-8                                                  554.1Ki ±   1%                554.2Ki ±   1%       ~ (p=0.937 n=6)
RangeQuery/expr=stdvar_over_time(a_one[1d]),steps=1000-8                                                 588.4Ki ±   6%                589.7Ki ±   0%       ~ (p=0.240 n=6)
RangeQuery/expr=stdvar_over_time(a_ten[1d]),steps=1-8                                                    756.0Ki ±   1%                753.2Ki ±   1%       ~ (p=0.394 n=6)
RangeQuery/expr=stdvar_over_time(a_ten[1d]),steps=100-8                                                  789.6Ki ±   5%                789.6Ki ±   5%       ~ (p=1.000 n=6)
RangeQuery/expr=stdvar_over_time(a_ten[1d]),steps=1000-8                                                1013.3Ki ±  23%               1014.5Ki ±  23%       ~ (p=0.773 n=6)
RangeQuery/expr=stdvar_over_time(a_hundred[1d]),steps=1-8                                                2.658Mi ±   1%                2.659Mi ±   1%       ~ (p=0.093 n=6)
RangeQuery/expr=stdvar_over_time(a_hundred[1d]),steps=100-8                                              2.790Mi ±   4%                2.670Mi ±   9%       ~ (p=0.937 n=6)
RangeQuery/expr=stdvar_over_time(a_hundred[1d]),steps=1000-8                                             3.332Mi ± 132%                3.335Mi ± 132%       ~ (p=0.619 n=6)
RangeQuery/expr=histogram_stddev(rate(h_one[5m])),steps=1-8                                              20.94Ki ±   0%                20.94Ki ±   0%       ~ (p=0.643 n=6)
RangeQuery/expr=histogram_stddev(rate(h_one[5m])),steps=100-8                                            26.27Ki ±   0%                26.27Ki ±   0%       ~ (p=0.660 n=6)
RangeQuery/expr=histogram_stddev(rate(h_one[5m])),steps=1000-8                                           75.79Ki ±   0%                75.74Ki ±   0%       ~ (p=0.063 n=6)
RangeQuery/expr=histogram_stddev(rate(h_ten[5m])),steps=1-8                                              138.6Ki ±   0%                138.6Ki ±   0%       ~ (p=0.132 n=6)
RangeQuery/expr=histogram_stddev(rate(h_ten[5m])),steps=100-8                                            170.3Ki ±   0%                170.3Ki ±   0%  -0.03% (p=0.026 n=6)
RangeQuery/expr=histogram_stddev(rate(h_ten[5m])),steps=1000-8                                           464.8Ki ±   0%                464.8Ki ±   0%       ~ (p=0.818 n=6)
RangeQuery/expr=histogram_stddev(rate(h_hundred[5m])),steps=1-8                                          1.227Mi ±   0%                1.226Mi ±   0%       ~ (p=0.180 n=6)
RangeQuery/expr=histogram_stddev(rate(h_hundred[5m])),steps=100-8                                        1.509Mi ±   0%                1.506Mi ±   0%       ~ (p=0.240 n=6)
RangeQuery/expr=histogram_stddev(rate(h_hundred[5m])),steps=1000-8                                       4.122Mi ±   0%                4.122Mi ±   0%       ~ (p=0.937 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_one[5m])),steps=1-8                                              20.94Ki ±   0%                20.94Ki ±   0%       ~ (p=0.998 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_one[5m])),steps=100-8                                            26.27Ki ±   0%                26.27Ki ±   0%       ~ (p=1.000 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_one[5m])),steps=1000-8                                           75.76Ki ±   0%                75.76Ki ±   0%       ~ (p=0.563 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_ten[5m])),steps=1-8                                              138.6Ki ±   0%                138.6Ki ±   0%       ~ (p=0.619 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_ten[5m])),steps=100-8                                            170.3Ki ±   0%                170.2Ki ±   0%  -0.03% (p=0.004 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_ten[5m])),steps=1000-8                                           465.0Ki ±   0%                464.6Ki ±   0%  -0.09% (p=0.015 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_hundred[5m])),steps=1-8                                          1.228Mi ±   0%                1.226Mi ±   0%  -0.14% (p=0.004 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_hundred[5m])),steps=100-8                                        1.509Mi ±   0%                1.506Mi ±   0%  -0.22% (p=0.041 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_hundred[5m])),steps=1000-8                                       4.118Mi ±   0%                4.121Mi ±   0%       ~ (p=0.485 n=6)
geomean                                                                                                  531.0Ki                       528.6Ki         -0.44%

                                                                   │ /Users/neerajgartia/Desktop/dev/cp/before_1000.txt │ /Users/neerajgartia/Desktop/dev/cp/after_1000.txt │
                                                                   │                     allocs/op                      │         allocs/op          vs base                │
RangeQuery/expr=stddev_over_time(a_one[1d]),steps=1-8                                                        396.0 ± 0%                  396.0 ± 0%       ~ (p=1.000 n=6) ¹
RangeQuery/expr=stddev_over_time(a_one[1d]),steps=100-8                                                      398.0 ± 0%                  398.0 ± 0%       ~ (p=1.000 n=6)
RangeQuery/expr=stddev_over_time(a_one[1d]),steps=1000-8                                                     422.0 ± 0%                  422.0 ± 0%       ~ (p=0.758 n=6)
RangeQuery/expr=stddev_over_time(a_ten[1d]),steps=1-8                                                       2.445k ± 0%                 2.445k ± 0%       ~ (p=0.905 n=6)
RangeQuery/expr=stddev_over_time(a_ten[1d]),steps=100-8                                                     2.475k ± 0%                 2.475k ± 0%       ~ (p=0.939 n=6)
RangeQuery/expr=stddev_over_time(a_ten[1d]),steps=1000-8                                                    2.701k ± 1%                 2.703k ± 2%       ~ (p=0.913 n=6)
RangeQuery/expr=stddev_over_time(a_hundred[1d]),steps=1-8                                                   22.30k ± 0%                 22.31k ± 0%       ~ (p=0.247 n=6)
RangeQuery/expr=stddev_over_time(a_hundred[1d]),steps=100-8                                                 22.61k ± 0%                 22.62k ± 3%  +0.02% (p=0.024 n=6)
RangeQuery/expr=stddev_over_time(a_hundred[1d]),steps=1000-8                                                24.69k ± 2%                 24.69k ± 2%       ~ (p=0.554 n=6)
RangeQuery/expr=stdvar_over_time(a_one[1d]),steps=1-8                                                        396.0 ± 0%                  396.0 ± 0%       ~ (p=1.000 n=6) ¹
RangeQuery/expr=stdvar_over_time(a_one[1d]),steps=100-8                                                      398.0 ± 0%                  398.0 ± 0%       ~ (p=1.000 n=6)
RangeQuery/expr=stdvar_over_time(a_one[1d]),steps=1000-8                                                     422.0 ± 0%                  422.0 ± 0%       ~ (p=1.000 n=6)
RangeQuery/expr=stdvar_over_time(a_ten[1d]),steps=1-8                                                       2.446k ± 0%                 2.445k ± 0%       ~ (p=0.799 n=6)
RangeQuery/expr=stdvar_over_time(a_ten[1d]),steps=100-8                                                     2.475k ± 0%                 2.475k ± 0%       ~ (p=1.000 n=6)
RangeQuery/expr=stdvar_over_time(a_ten[1d]),steps=1000-8                                                    2.701k ± 0%                 2.701k ± 0%       ~ (p=0.437 n=6)
RangeQuery/expr=stdvar_over_time(a_hundred[1d]),steps=1-8                                                   22.31k ± 0%                 22.31k ± 0%       ~ (p=0.323 n=6)
RangeQuery/expr=stdvar_over_time(a_hundred[1d]),steps=100-8                                                 22.61k ± 0%                 22.61k ± 0%       ~ (p=0.985 n=6)
RangeQuery/expr=stdvar_over_time(a_hundred[1d]),steps=1000-8                                                24.68k ± 2%                 24.68k ± 2%       ~ (p=0.853 n=6)
RangeQuery/expr=histogram_stddev(rate(h_one[5m])),steps=1-8                                                  316.0 ± 0%                  316.0 ± 0%       ~ (p=1.000 n=6) ¹
RangeQuery/expr=histogram_stddev(rate(h_one[5m])),steps=100-8                                                449.0 ± 0%                  449.0 ± 0%       ~ (p=1.000 n=6) ¹
RangeQuery/expr=histogram_stddev(rate(h_one[5m])),steps=1000-8                                              1.627k ± 0%                 1.627k ± 0%       ~ (p=1.000 n=6) ¹
RangeQuery/expr=histogram_stddev(rate(h_ten[5m])),steps=1-8                                                 1.411k ± 0%                 1.411k ± 0%       ~ (p=1.000 n=6)
RangeQuery/expr=histogram_stddev(rate(h_ten[5m])),steps=100-8                                               1.841k ± 0%                 1.840k ± 0%       ~ (p=0.242 n=6)
RangeQuery/expr=histogram_stddev(rate(h_ten[5m])),steps=1000-8                                              5.495k ± 0%                 5.495k ± 0%       ~ (p=1.000 n=6)
RangeQuery/expr=histogram_stddev(rate(h_hundred[5m])),steps=1-8                                             12.31k ± 0%                 12.31k ± 0%       ~ (p=1.000 n=6)
RangeQuery/expr=histogram_stddev(rate(h_hundred[5m])),steps=100-8                                           15.62k ± 0%                 15.62k ± 0%       ~ (p=0.188 n=6)
RangeQuery/expr=histogram_stddev(rate(h_hundred[5m])),steps=1000-8                                          43.26k ± 0%                 43.26k ± 0%       ~ (p=0.554 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_one[5m])),steps=1-8                                                  316.0 ± 0%                  316.0 ± 0%       ~ (p=1.000 n=6) ¹
RangeQuery/expr=histogram_stdvar(rate(h_one[5m])),steps=100-8                                                449.0 ± 0%                  449.0 ± 0%       ~ (p=1.000 n=6) ¹
RangeQuery/expr=histogram_stdvar(rate(h_one[5m])),steps=1000-8                                              1.627k ± 0%                 1.627k ± 0%       ~ (p=1.000 n=6) ¹
RangeQuery/expr=histogram_stdvar(rate(h_ten[5m])),steps=1-8                                                 1.411k ± 0%                 1.411k ± 0%       ~ (p=1.000 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_ten[5m])),steps=100-8                                               1.841k ± 0%                 1.841k ± 0%       ~ (p=0.182 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_ten[5m])),steps=1000-8                                              5.495k ± 0%                 5.494k ± 0%       ~ (p=0.697 n=6)
RangeQuery/expr=histogram_stdvar(rate(h_hundred[5m])),steps=1-8                                             12.31k ± 0%                 12.31k ± 0%       ~ (p=0.455 n=6)
…